% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UVdose_datacheck.R
\name{uv_data_check}
\alias{uv_data_check}
\title{Data check for UV dose calculation}
\usage{
uv_data_check(data)
}
\arguments{
\item{data}{Input dataframe to be used for UV dose calculation}
}
\value{
A message checking validity of input data

None, returns a message about input data validity.
}
\description{
UV dose calculation requires date, latitude, and longitude inputs.
}
\examples{
mysample <- data.frame(id = c("id000016", "id000021"),
      date = as.Date(c("2009-05-15", "2008-08-04")),
      easting = c(519000, 365000),
      northing = c(176000, 172000))
uv_data_check(mysample)
}
